Move schema files (#2919)
