+++ tests/schema/test_validate_schema.py
@@ -141,20 +141,20 @@ class TestValidateUnifiedSchema(TestCase):
         assert len(UNIFIED_SCHEMA["properties"]["Resources"]["additionalProperties"]["anyOf"]) > 1000
         assert (
             "The set of properties must conform to the defined `Type`"
-            in UNIFIED_SCHEMA["definitions"]["schema_source__aws_serverless_statemachine__ApiEvent"]["properties"][
-                "Properties"
-            ]["markdownDescription"]
+            in UNIFIED_SCHEMA["definitions"][
+                "samtranslator__internal__schema_source__aws_serverless_statemachine__ApiEvent"
+            ]["properties"]["Properties"]["markdownDescription"]
         )
 
         # Contains all definitions from SAM-only schema (except rule that ignores non-SAM)
         sam_defs = copy.deepcopy(SCHEMA["definitions"])
-        del sam_defs["schema_source__any_cfn_resource__Resource"]
+        del sam_defs["samtranslator__internal__schema_source__any_cfn_resource__Resource"]
         assert sam_defs.items() <= UNIFIED_SCHEMA["definitions"].items()
 
         # Contains all resources from SAM-only schema (except rule that ignores non-SAM)
         unified_resources = UNIFIED_SCHEMA["properties"]["Resources"]["additionalProperties"]["anyOf"]
         for v in SCHEMA["properties"]["Resources"]["additionalProperties"]["anyOf"]:
-            if v["$ref"] != "#/definitions/schema_source__any_cfn_resource__Resource":
+            if v["$ref"] != "#/definitions/samtranslator__internal__schema_source__any_cfn_resource__Resource":
                 assert v in unified_resources
 
     @parameterized.expand(
